make dedicated FAQs ui component
